Move App Password package to Site Credentials
