+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/login/sitecredentials/applicationpassword/ApplicationPasswordTutorialFragment.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.login.applicationpassword
+package com.woocommerce.android.ui.login.sitecredentials.applicationpassword
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -11,8 +11,8 @@ import com.woocommerce.android.support.help.HelpOrigin
 import com.woocommerce.android.support.requests.SupportRequestFormActivity
 import com.woocommerce.android.ui.base.BaseFragment
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
-import com.woocommerce.android.ui.login.applicationpassword.ApplicationPasswordTutorialViewModel.OnContactSupport
-import com.woocommerce.android.ui.login.applicationpassword.ApplicationPasswordTutorialViewModel.OnContinue
+import com.woocommerce.android.ui.login.sitecredentials.applicationpassword.ApplicationPasswordTutorialViewModel.OnContactSupport
+import com.woocommerce.android.ui.login.sitecredentials.applicationpassword.ApplicationPasswordTutorialViewModel.OnContinue
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
